letter-combinations-of-a-phone-number/lettercombinationsofaphonenumber.go
package lettercombinationsofaphonenumber

var digitToLettersMap = map[string][]string{
	"2": []string{"a", "b", "c"},
	"3": []string{"d", "e", "f"},
	"4": []string{"g", "h", "i"},
	"5": []string{"j", "k", "l"},
	"6": []string{"m", "n", "o"},
	"7": []string{"p", "q", "r", "s"},
	"8": []string{"t", "u", "v"},
	"9": []string{"w", "x", "y", "z"},
}

func letterCombinations(digits string) []string {
	if letters, isValidDigit := digitToLettersMap[digits]; isValidDigit {
		return letters
	}
	return nil
}
